backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
geoalchemy2
psycopg2-binary
pydantic
pydantic-settings
celery
redis
python-dotenv
httpx
numpy
pandas
scikit-learn
torch
scipy
dowhy
statsmodels
prophet
h3
python-multipart
websockets
